XD-49 - Move Redis Queue Channel Adapters to SI
