+++ modules/source/tap.xml
@@ -18,13 +18,13 @@
 
 	<int:channel id="output"/>
 
-	<bean id="channelRegistry" class="org.springframework.integration.x.redis.RedisChannelRegistry">
+	<bean id="channelRegistry" class="org.springframework.integration.redis.channel.registry.RedisChannelRegistry">
 		<constructor-arg ref="redisConnectionFactory"/>
 	</bean>
 
 	<bean id="connectionFactory" class="org.springframework.data.redis.connection.lettuce.LettuceConnectionFactory">
 		<constructor-arg index="0" value="${redis.hostname:localhost}"/>
-		<constructor-arg index="1" value="${redis.port:6379}"/>	  
+		<constructor-arg index="1" value="${redis.port:6379}"/>
 	</bean>
 
 </beans>
